Split siphon into multiple parts, the main, the stream, and the pty. Add my Vagrantfile to be able to launch the vm to create it
