+++ .gitignore
@@ -12,3 +12,4 @@ Makefile.in
 /install-sh
 /missing
 .deps
+.vagrant
