+++ configure.ac
@@ -11,32 +11,7 @@ AC_PROG_CC
 
 # Checks for libraries.
 
-#AC_CHECK_LIB(m, acos,
-#  [],
-#  [AC_MSG_ERROR([m libraries not found.])]
-#)
-#
-#AC_CHECK_LIB(tcl8.6, Tcl_Access,
-#  [],
-#  [AC_MSG_ERROR([tcl libraries not found.])]
-#)
-#
-#AC_CHECK_LIB(expect, exp_spawnv,
-#  [],
-#  [AC_MSG_ERROR([expect libraries not found.])]
-#)
-
-#AC_CHECK_LIB(util, openpty,
-#  [],
-#  [AC_MSG_ERROR([util libraries not found.])]
-#)
-
 # Checks for header files.
-
-# AC_CHECK_HEADERS(tcl8.6/expect.h,
-#   [],
-#   [AC_MSG_ERROR([expect header files not found.]); break]
-# )
 
 # Checks for typedefs, structures, and compiler characteristics.
