+++ src/Makefile.am
@@ -1,2 +1,2 @@
 bin_PROGRAMS = siphon
-siphon_SOURCES = siphon.c
+siphon_SOURCES = siphon.c siphon_pty.c siphon_stream.c
