+++ src/siphon_pty.h
@@ -1,0 +1,1 @@
+int exp_spawnv(char *prefix,bool prefix_set,char *file, char *argv[]);
